refactor: address PR review on policy re-acceptance modal
- Rely on Modal defaults instead of a custom size className.
- Use the Surface component (filled) for the "What's changed" card.
- Drop the extra wrapper so the checkbox row and button use the
  ModalBody gap directly.
- Use the @op/ui Button (link variant) for the back arrow and the
  in-text policy links instead of raw <button> elements; moving the
  links beside the checkbox rather than inside its label removes the
  need for manual press-propagation handling.
- Colocate the single-use eligibility helper next to the component
  (eligibility.ts) instead of a standalone utils file. It stays a pure
  module so its unit test can import it without pulling the client
  component (and next-intl) into the node test env.
- Add the new strings to every translation dictionary (ar, bn, fr, pt,
  so) in addition to en/es.
