+++ apps/app/src/lib/i18n/dictionaries/ar.json
@@ -35,6 +35,14 @@
   "Exit": "خروج",
   "Effective Date: March 15, 2025": "تاريخ السريان: 15 مارس 2025",
   "Privacy Policy": "سياسة الخصوصية",
+  "Terms of Use": "شروط الاستخدام",
+  "Code of Conduct": "مدونة السلوك",
+  "We've updated our policies.": "لقد قمنا بتحديث سياساتنا.",
+  "Review the changes and accept to keep using Common.": "راجع التغييرات ووافق عليها لمواصلة استخدام Common.",
+  "What's changed": "ما الذي تغيّر",
+  "Common now works with a third-party service that automatically reviews content posted on the platform to keep the community safe and uphold our Code of Conduct. Our Terms of Use and Privacy Policy now explain how this works and what it means for your data.": "أصبح Common الآن يعمل مع خدمة تابعة لجهة خارجية تراجع تلقائيًا المحتوى المنشور على المنصة للحفاظ على أمان المجتمع ودعم مدونة السلوك الخاصة بنا. توضّح الآن شروط الاستخدام وسياسة الخصوصية كيفية عمل ذلك وما يعنيه بالنسبة لبياناتك.",
+  "I have read and agree to the <terms>Terms of Use</terms>, <privacy>Privacy Policy</privacy>, and <conduct>Code of Conduct</conduct>.": "لقد قرأت وأوافق على <terms>شروط الاستخدام</terms> و<privacy>سياسة الخصوصية</privacy> و<conduct>مدونة السلوك</conduct>.",
+  "Agree and continue": "الموافقة والمتابعة",
   "Media": "الوسائط",
   "We've found your organization": "لقد وجدنا مؤسستك",
   "join_subheader": "بناءً على نطاق بريدك الإلكتروني، يمكنك الانضمام إلى هذه المؤسسة.",
