+++ apps/app/src/lib/i18n/dictionaries/bn.json
@@ -35,6 +35,14 @@
   "Exit": "প্রস্থান",
   "Effective Date: March 15, 2025": "কার্যকর হওয়ার তারিখ: ১৫ মার্চ, ২০২৫",
   "Privacy Policy": "গোপনীয়তা নীতি",
+  "Terms of Use": "ব্যবহারের শর্তাবলী",
+  "Code of Conduct": "আচরণবিধি",
+  "We've updated our policies.": "আমরা আমাদের নীতিমালা আপডেট করেছি।",
+  "Review the changes and accept to keep using Common.": "পরিবর্তনগুলো পর্যালোচনা করুন এবং Common ব্যবহার চালিয়ে যেতে সম্মতি দিন।",
+  "What's changed": "যা পরিবর্তিত হয়েছে",
+  "Common now works with a third-party service that automatically reviews content posted on the platform to keep the community safe and uphold our Code of Conduct. Our Terms of Use and Privacy Policy now explain how this works and what it means for your data.": "Common এখন একটি তৃতীয় পক্ষের পরিষেবার সাথে কাজ করে যা সম্প্রদায়কে নিরাপদ রাখতে এবং আমাদের আচরণবিধি বজায় রাখতে প্ল্যাটফর্মে পোস্ট করা কনটেন্ট স্বয়ংক্রিয়ভাবে পর্যালোচনা করে। আমাদের ব্যবহারের শর্তাবলী এবং গোপনীয়তা নীতি এখন ব্যাখ্যা করে এটি কীভাবে কাজ করে এবং আপনার ডেটার জন্য এর অর্থ কী।",
+  "I have read and agree to the <terms>Terms of Use</terms>, <privacy>Privacy Policy</privacy>, and <conduct>Code of Conduct</conduct>.": "আমি <terms>ব্যবহারের শর্তাবলী</terms>, <privacy>গোপনীয়তা নীতি</privacy> এবং <conduct>আচরণবিধি</conduct> পড়েছি ও সম্মত হয়েছি।",
+  "Agree and continue": "সম্মত হয়ে চালিয়ে যান",
   "Media": "মিডিয়া",
   "We've found your organization": "আমরা আপনার সংস্থা খুঁজে পেয়েছি",
   "join_subheader": "আপনার ইমেইল ডোমেইনের ভিত্তিতে, আপনার এই সংস্থায় যোগ দেওয়ার অ্যাক্সেস আছে।",
