+++ apps/app/src/lib/i18n/dictionaries/fr.json
@@ -35,6 +35,14 @@
   "Exit": "Quitter",
   "Effective Date: March 15, 2025": "Date d'entrée en vigueur : 15 mars 2025",
   "Privacy Policy": "Politique de confidentialité",
+  "Terms of Use": "Conditions d'utilisation",
+  "Code of Conduct": "Code de conduite",
+  "We've updated our policies.": "Nous avons mis à jour nos politiques.",
+  "Review the changes and accept to keep using Common.": "Consultez les changements et acceptez-les pour continuer à utiliser Common.",
+  "What's changed": "Ce qui a changé",
+  "Common now works with a third-party service that automatically reviews content posted on the platform to keep the community safe and uphold our Code of Conduct. Our Terms of Use and Privacy Policy now explain how this works and what it means for your data.": "Common travaille désormais avec un service tiers qui examine automatiquement le contenu publié sur la plateforme afin de préserver la sécurité de la communauté et de faire respecter notre Code de conduite. Nos Conditions d'utilisation et notre Politique de confidentialité expliquent maintenant comment cela fonctionne et ce que cela signifie pour vos données.",
+  "I have read and agree to the <terms>Terms of Use</terms>, <privacy>Privacy Policy</privacy>, and <conduct>Code of Conduct</conduct>.": "J'ai lu et j'accepte les <terms>Conditions d'utilisation</terms>, la <privacy>Politique de confidentialité</privacy> et le <conduct>Code de conduite</conduct>.",
+  "Agree and continue": "Accepter et continuer",
   "Media": "Moyens de communication",
   "We've found your organization": "Nous avons trouvé votre organisation",
   "join_subheader": "En fonction du domaine de votre courriel, vous avez accès pour rejoindre cette organisation.",
