+++ apps/app/src/lib/i18n/dictionaries/pt.json
@@ -35,6 +35,14 @@
   "Exit": "Sair",
   "Effective Date: March 15, 2025": "Data de Vigência: 15 de março de 2025",
   "Privacy Policy": "Política de Privacidade",
+  "Terms of Use": "Termos de Uso",
+  "Code of Conduct": "Código de Conduta",
+  "We've updated our policies.": "Atualizamos nossas políticas.",
+  "Review the changes and accept to keep using Common.": "Revise as mudanças e aceite para continuar usando o Common.",
+  "What's changed": "O que mudou",
+  "Common now works with a third-party service that automatically reviews content posted on the platform to keep the community safe and uphold our Code of Conduct. Our Terms of Use and Privacy Policy now explain how this works and what it means for your data.": "O Common agora trabalha com um serviço de terceiros que revisa automaticamente o conteúdo publicado na plataforma para manter a comunidade segura e defender nosso Código de Conduta. Nossos Termos de Uso e Política de Privacidade agora explicam como isso funciona e o que significa para os seus dados.",
+  "I have read and agree to the <terms>Terms of Use</terms>, <privacy>Privacy Policy</privacy>, and <conduct>Code of Conduct</conduct>.": "Li e concordo com os <terms>Termos de Uso</terms>, a <privacy>Política de Privacidade</privacy> e o <conduct>Código de Conduta</conduct>.",
+  "Agree and continue": "Concordar e continuar",
   "Media": "Mídia",
   "We've found your organization": "Encontramos sua organização",
   "join_subheader": "Com base no domínio do seu e-mail, você tem acesso para se juntar a esta organização.",
