+++ apps/app/src/lib/i18n/dictionaries/so.json
@@ -35,6 +35,14 @@
   "Exit": "Ka bax",
   "Effective Date: March 15, 2025": "Taariikhda Faa'iidada: Maarso 15, 2025",
   "Privacy Policy": "Xeerka Asturnaanta",
+  "Terms of Use": "Shuruudaha Isticmaalka",
+  "Code of Conduct": "Xeerka Anshaxa",
+  "We've updated our policies.": "Waan cusbooneysiinnay siyaasadahayaga.",
+  "Review the changes and accept to keep using Common.": "Dib u eeg isbeddellada oo aqbal si aad u sii isticmaasho Common.",
+  "What's changed": "Waxa isbeddelay",
+  "Common now works with a third-party service that automatically reviews content posted on the platform to keep the community safe and uphold our Code of Conduct. Our Terms of Use and Privacy Policy now explain how this works and what it means for your data.": "Common hadda wuxuu la shaqeeyaa adeeg saddexaad oo si toos ah u eega waxa la soo dhigo plat-foomka si loo ilaaliyo bulshada iyo in la ilaaliyo Xeerka Anshaxa. Shuruudaha Isticmaalka iyo Xeerka Asturnaanta hadda waxay sharxayaan sida ay taasi u shaqeyso iyo waxa ay ka dhigan tahay xogtaada.",
+  "I have read and agree to the <terms>Terms of Use</terms>, <privacy>Privacy Policy</privacy>, and <conduct>Code of Conduct</conduct>.": "Waan akhriyay oo waan aqbalay <terms>Shuruudaha Isticmaalka</terms>, <privacy>Xeerka Asturnaanta</privacy>, iyo <conduct>Xeerka Anshaxa</conduct>.",
+  "Agree and continue": "Aqbal oo sii wad",
   "Media": "Warbaahinta",
   "We've found your organization": "Waxaan helay ururkaaga",
   "join_subheader": "Iyadoo lagu saleynayo domain-ka iimeelkaaga, waxaad xaq u leedahay inaad ku biirto ururkan.",
